# services/business-agency-saas-api/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  /api/ai-agents:
    post:
      summary: 'Store a new agent.'
      operationId: storeANewAgent
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                slug:
                  type: string
                  description: 'Ensure slug is unique per tenant. Must contain only letters, numbers, dashes and underscores. Must not be greater than 255 characters.'
                  example: 'n'
                brain:
                  type: string
                  description: ''
                  example: architecto
                model:
                  type: string
                  description: "The service key (e.g. 'openai')."
                  example: architecto
                system_prompt:
                  type: string
                  description: "e.g. 'gpt-4'."
                  example: architecto
                  nullable: true
                user_prompt:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                tools:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                triggers:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                is_active:
                  type: boolean
                  description: 'Logic to save triggers handled below or via relationship.'
                  example: true
              required:
                - name
                - slug
                - brain
                - model
  '/api/ai-agents/{id}':
    put:
      summary: 'Update an existing agent.'
      operationId: updateAnExistingAgent
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                slug:
                  type: string
                  description: 'Must contain only letters, numbers, dashes and underscores. Must not be greater than 255 characters.'
                  example: 'n'
                brain:
                  type: string
                  description: ''
                  example: architecto
                model:
                  type: string
                  description: ''
                  example: architecto
                system_prompt:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                user_prompt:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                tools:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                triggers:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                is_active:
                  type: boolean
                  description: ''
                  example: true
    delete:
      summary: 'Delete an agent.'
      operationId: deleteAnAgent
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the ai agent.'
        example: architecto
        required: true
        schema:
          type: string
...